Add files via upload
Open Source Assignment
 

Submitted to: Miss. Milanjit Kaur


Submitted by :
Name – Abhishek Sahore
Registration no. – 11611735
Roll no. – A27


Problem Statement

Create a login and signup page as below. Save the entries in the database. When you are done with signup, it should be directed to log in up. 

Description

I have created login page, signup page and a home page using HTML5 and CSS3 and I have written its backend in PHP and also used MySQL for the database.

Signup page
The Signup page takes input as follow:
1-	Name 
2-	Username
3-	Registration ID 
4-	Mobile Number
5-	Email ID
6-	Password 
7-	Confirm Password
If the signup form is empty, the form will not be submitted and the form data will not go to the database. The form will only get submitted if all the form fields will be filled. 
If the any of the data filled in the form input fields like username, registrationID, mobile, email turns out to be duplicate or pre-exists in the database, the form will not get submitted as this data is unique to the specific user.
If the password and confirm password does not match, then also the form will not get submitted.
The Signup page has a Signup button and a link to the Login page in case if the user already exists.

Login page
The Login page takes input as follow:
1-	Username 
2-	Password
The Login page will only be submitted to the database if both the fields are filled. 
 If the username does not exist, it will ask the user to fill the correct username or Signup by quoting “User does not exist. Signup or enter valid username”.
If the user enters the correct username but invalid password, then it will ask the user to fill the correct password.
If both the credentials are correct, then it will take the user to the home page of this assignment.

Home page
Home page contains the following content:
1-	“Welcome, “ string followed by the username.
2-	Logout link button which will take the user to the login page again.


Database
The Database used in this assignment is the MySQL database. It has the following fields:
1-	id – it is an int type field which is of an auto_increment, primary key, and not null type.
2-	name -  it is a varchar field which is of not null type.
3-	username – it is a varchar type field which is of a not null and unique type.
4-	registrationId – it is a varchar field which is of a not null and unique type.
5-	email – it is a varchar field which is of a not null and unique type.
6-	mobile – it is a varchar field which is not null and unique type.
7-	Password – it is a varchar field which is of a not null type.


Link to my GitHub repository for this assignment:









Test Cases

Test Cases for Signup page
1-	If all the fields are empty or any of the fields is empty.

 
 



2-	If data filled in the input fields like username, registrationId, email, mobile already exists in the database.

 
 







3-	If the password and confirm password does not match.

 

 







4-	If all the credentials are valid according to the program.


 


 




Test cases for the Login page
1-	If the username and password both are invalid.

 

 





2-	If the username is valid but the password is invalid.

 



3-	If the form is empty.

 



4-	If the username and password both are valid, it will lead to the home page.

 

 


Test cases for the home page
1-	If the user clicked on the Logout link button, it will lead to the login page.

 


 




Screenshots of the MySQL Database